model2/granite/requirements.txt
streamlit
huggingface_hub
matplotlib
wordcloud
qrcode
pillow
PyPDF2 